Be consistent on naming. Simulation -> Runner
Simulations could either be seen as the propagation of one esnemble
member or as the simulation code that is executed somewhere propagating
multiple ensemble members. This is confusing. So now the simulation word
is exchanged by (Model Task) Runner everywhere where it annotates a
Simulation core running somewhere that is waiting for ensemble members
to propagate.

Further this commit cleans up a bit some comments