Link against pdaf wrapper functions dynamically
Thus one can replace the pdaf wrapper functions by user defined as
needed for the pure-parflow assimilation project. Unfortunately this is
kind of dirty from an software engineering point of view as there are
a lot of code duplications. Still it is really hard to define a flexible
and at the same time minimalistic interface.

Furthermore I do not know how much performance we loose by using shared
libraries and especially the preload trick.