Implement an API that supports the hidden state
The hidden state is saved in an extra field that may not be touched by
the do_update_step function of the assimilator. The correct hidden state
(if one on field registration was given) will be always linked to the
corresponding state vector which is assimilated. When e.g. propagating a
state vector the corresponding hidden state is also sent to the runner.

As the hidden state is not changed by the update step it is enough to
save it once (and not split into two vectors: background and analysis
state as needed for the state vector)