Add max_devel_jobs option for SlurmJuwelsCluster
On the Juwels cluster only some jobs can be run in the devel partition.
The devel partition is most of the time empty so its a good idea to
start the first jobs in there so there will be something running. Other
Runners are started in the standard partition to have even more runners.