README: list licenses of dependencies
Many open source software licenses (e.g., MPL2.0, LGPLv3, 3-clause BSD)
demand that if a piece of software is used, then it must be named as a
dependency, its license must be indicated, _and_ a license copy must be
made available.

This commit adds the license texts and updates the README accordingly.