Replace killall by pkill and wait before kill
killall is not installed on the ubuntu gitlab runner.

Sometimes pkill got called to early, especially when mpiexec was still
about to spawn some pocesses. This should now be fixed (in the cases
where it appeared) by adding a timeout. On supercomputers where one is
not relyiing on the systemms process handling but on the batch scheduler
this issue should be no problem.