Use char internally to denser pack data
WRF works on arrays in many different data types
(double/float/int/char/...) To migrate WRF states, those need to be
serialized and pumped through melissa. If we would transform everything to
double the amount of sent data would rathly double. With this commit we introduce
char to internally handle the data bytewise. This is a step to maybe later support type
invaraince using PDI.

Note:
- Index maps still need to know how dense they
  shall index the data (e.g. one index to every 4 bytes if indexing double
  arrays... , see bytes_per_element)
- the whole index map uses the same bytes_per_element
- only hidden and assimilated states can have different bytes_per_element
  variables
- n to m flows happen element wise (if bytes_per_element=4 it will e.g.
  only split after a multiple of 4 bytes)