Avoid wrongly triggered ADD_RUNNER events on server side
ADD_RUNNER events now are only triggeered if a runner is really ready to
perform work on the server side and not already when the runners first
contacts the server to retrieve nxm connection information. This avoids
a wrong runner count on the server side when a runner crashes before it
requested work from the server for the first time. (The server will
never recognize crashing runners if they never had work to do. In such
cases crashing runners are caught by the launcher only)