Launcher: minimize signal handler actions
GDB showed Python processes waiting for semaphore locks with the
following (truncated) backtrace:
* PyThread_acquire_lock_timed
* lock_PyThread_acquire_lock
* PyEval_EvalFrameEx (/usr/lib/python3.8/subprocess.py, in _wait (self=<Popen(_waitpid_lock=..., args=['/usr/bin/mpiexec', ...])>, timeout=None))
* PyEval_EvalFrameEx (melissa-da/launcher/cluster/local_cluster.py, in CleanUp (self=<LocalCluster(jobs={1099: <Popen(_waitpid_lock=<>, ...)>})>))
* _PyEval_EvalCodeWithName (name='signal_handler', qualname='run_melissa_da_study.<locals>.signal_handler')

The `run_melissa_da_study` signal handler attempted to terminate all
child processes and since it was called asynchronously, it can obviously
be a source of concurreny problems. This commit is based on the
observation that setting the variable `running` to false is all that is
needed to make `run_melissa_da_study` terminate all child processes and
return to the caller.